+++ integration/spark-common/src/main/scala/org/apache/carbondata/spark/util/CommonUtil.scala
@@ -345,11 +345,4 @@ object CommonUtil {
 
     csvColumns
   }
-
-  def validateColumnNames(columnName: String, columnNameCopy: String): Unit = {
-    if (!columnName.equalsIgnoreCase(columnNameCopy)) {
-      throw new MalformedCarbonCommandException(
-        "Column names provided are different. Both the column names should be same")
-    }
-  }
 }
