+++ integration/spark-common/src/main/scala/org/apache/carbondata/spark/util/DataTypeConverterUtil.scala
@@ -18,6 +18,7 @@
 package org.apache.carbondata.spark.util
 
 import org.apache.carbondata.core.metadata.datatype.DataType
+import org.apache.carbondata.format.{DataType => ThriftDataType}
 
 object DataTypeConverterUtil {
   def convertToCarbonType(dataType: String): DataType = {
@@ -89,33 +90,22 @@ object DataTypeConverterUtil {
    * @param dataType
    * @return
    */
-  def convertToThriftDataType(dataType: String): org.apache.carbondata.format.DataType = {
+  def convertToThriftDataType(dataType: String): ThriftDataType = {
     if (null == dataType) {
       return null
     }
     dataType match {
-      case "string" =>
-        org.apache.carbondata.format.DataType.STRING
-      case "int" =>
-        org.apache.carbondata.format.DataType.INT
-      case "short" =>
-        org.apache.carbondata.format.DataType.SHORT
-      case "long" | "bigint" =>
-        org.apache.carbondata.format.DataType.LONG
-      case "double" =>
-        org.apache.carbondata.format.DataType.DOUBLE
-      case "decimal" =>
-        org.apache.carbondata.format.DataType.DECIMAL
-      case "date" =>
-        org.apache.carbondata.format.DataType.DATE
-      case "timestamp" =>
-        org.apache.carbondata.format.DataType.TIMESTAMP
-      case "array" =>
-        org.apache.carbondata.format.DataType.ARRAY
-      case "struct" =>
-        org.apache.carbondata.format.DataType.STRUCT
-      case _ =>
-        org.apache.carbondata.format.DataType.STRING
+      case "string" => ThriftDataType.STRING
+      case "int" => ThriftDataType.INT
+      case "short" => ThriftDataType.SHORT
+      case "long" | "bigint" => ThriftDataType.LONG
+      case "double" => ThriftDataType.DOUBLE
+      case "decimal" => ThriftDataType.DECIMAL
+      case "date" => ThriftDataType.DATE
+      case "timestamp" => ThriftDataType.TIMESTAMP
+      case "array" => ThriftDataType.ARRAY
+      case "struct" => ThriftDataType.STRUCT
+      case _ => ThriftDataType.STRING
     }
   }
 }
